Crap 2.3

+++ USBLMulti/source/disc.c
@@ -111,7 +111,7 @@ void __Disc_SetVMode(void)
 void __Disc_SetTime(void)
 {
 	/* Extern */
-	extern void settime(long long);
+	extern void settime(u64);
 
 	/* Set proper time */
 	settime(secs_to_ticks(time(NULL) - 946684800));
